Add memcache to the nearest API.
It truncates the incoming location to 4 decimal point in lat, and 5 decimal points in lon. This results in the world be split into ~1m * ~4m rectangles, with all points within a rectangle being treated the same.

This should be more than enough accuracy for this application, and makes it easy to cache geospacial results. Additionaly as the API return the exact location of the nearest bus stop, the client can calculate a more accurate distance if it wants.