Add filtering by day and departure roll around.
We now filter the departure times by their valid days so that results for a weekday don't appear on the week end.

Also setup the roll around so that if there are no results for today (after the last bus of the day), we will get the first result for tomorrow.