Change how markers are stored
Now markers are stored on the map they are set on. This way it makes it possible to easily set the same marker on multiple maps.

Also the set marker functions now require you to pass the map you want the map to appear on. This helps with the point above.