Use background offsetting and change zoom
Offset the background so that the "center" of the map appears behind the logo. That way the visible portion of the background bears a resemblance to the area surrounding the user.

Also change the background zoom level logic so the background looks nicer. This is done by limiting the maximum amount of zoom the background is allowed (its currently set to the smallest scale that still displays building outlines). Finally the zoom difference has been reduced because I think it looks better.